Add files via upload
This code is meant as a visual aid for students in the CS 241 class I TA'd for. I decided to keep it all in one file and put the controls in a block comment at the top, that way I could send this code to a student who might be struggling and they could get it running with minimal hassle. Just copy and paste into a single file and hit play.